Fix incorrect behaviour in decrease_exp function
Change declaration of decrease_exp function: parameter (int reminder)
has changed to (bool has_reminder). Previous function gave incorrect
result on decimals like: 33586526831355.036500

Before:
33586526831355.036500 > 33586526831355.037
Now:
33586526831355.036500 >33586526831355.036

This change will also fix error in s21_div. In cases
when we need to shorten a decimal and in the end of it was:...25021...
Before:
...2
Now:
...3
Since there was 0 after 5 the program thought there was no reminder.

Change all the calls of decrease_exp function according to the new
declaration Add tests for s21_div with cases for this behaviour.